Cambiado el nombre del módulo y añadida la secuencia al tipo y al método create del pedido de venta
