+++ sale_order_types/views/sale_order_type_view.xml
@@ -2,9 +2,9 @@
 <openerp>
     <data>
 
-        <record id="sot_sale_order_typology_form_view" model="ir.ui.view">
-            <field name="name">sale.order.typology.form.view</field>
-            <field name="model">sale.order.typology</field>
+        <record id="sot_sale_order_type_form_view" model="ir.ui.view">
+            <field name="name">sale.order.type.form.view</field>
+            <field name="model">sale.order.type</field>
             <field name="arch" type="xml">
                 <form string="Type" version="7.0">
                     <sheet>
@@ -19,6 +19,7 @@
                         </group>
                         <group colspan="4">
                             <field name="warehouse_id"/>
+                            <field name="sequence_id"/>
                             <field name="description"/>
                         </group>
                     </sheet>
@@ -26,43 +27,44 @@
             </field>
         </record>
 
-        <record id="sot_sale_order_typology_tree_view" model="ir.ui.view">
-            <field name="name">sale.order.typology.tree.view</field>
-            <field name="model">sale.order.typology</field>
+        <record id="sot_sale_order_type_tree_view" model="ir.ui.view">
+            <field name="name">sale.order.type.tree.view</field>
+            <field name="model">sale.order.type</field>
             <field name="arch" type="xml">
                 <tree string="Type">
                     <field name="name"/>
                     <field name="warehouse_id"/>
+                    <field name="sequence_id"/>
                     <field name="description"/>
                 </tree>
             </field>
         </record>
 
-        <record id="sot_sale_order_typology_action" model="ir.actions.act_window">
+        <record id="sot_sale_order_type_action" model="ir.actions.act_window">
             <field name="name">Typology</field>
-            <field name="res_model">sale.order.typology</field>
+            <field name="res_model">sale.order.type</field>
             <field name="view_mode">form,tree</field>
-            <field name="view_id" ref="sot_sale_order_typology_tree_view"/>
+            <field name="view_id" ref="sot_sale_order_type_tree_view"/>
             <field name="help" type="html">
                 <p class="oe_view_nocontent_create">Click to set a new type for sale order.</p>
             </field>
         </record>
 
-        <record id="sot_sale_order_typology_action_form" model="ir.actions.act_window.view">
+        <record id="sot_sale_order_type_action_form" model="ir.actions.act_window.view">
             <field name="sequence" eval="2"/>
             <field name="view_mode">form</field>
-            <field name="view_id" ref="sot_sale_order_typology_form_view"/>
-            <field name="act_window_id" ref="sot_sale_order_typology_action"/>
+            <field name="view_id" ref="sot_sale_order_type_form_view"/>
+            <field name="act_window_id" ref="sot_sale_order_type_action"/>
         </record>
 
-        <record id="sot_sale_order_typology_action_tree" model="ir.actions.act_window.view">
+        <record id="sot_sale_order_type_action_tree" model="ir.actions.act_window.view">
             <field name="sequence" eval="1"/>
             <field name="view_mode">tree</field>
-            <field name="view_id" ref="sot_sale_order_typology_tree_view"/>
-            <field name="act_window_id" ref="sot_sale_order_typology_action"/>
+            <field name="view_id" ref="sot_sale_order_type_tree_view"/>
+            <field name="act_window_id" ref="sot_sale_order_type_action"/>
         </record>
 
-        <menuitem id="sot_sale_order_typology_menu" parent="base.menu_sales" action="sot_sale_order_typology_action" name="Typology" sequence="20"/>
+        <menuitem id="sot_sale_order_type_menu" parent="base.menu_sales" action="sot_sale_order_type_action" name="Typology" sequence="20"/>
 
     </data>
 </openerp>
