+++ sale_order_types/views/sale_order_view.xml
@@ -8,7 +8,7 @@
             <field name="inherit_id" ref="sale.view_order_form"/>
             <field name="arch" type="xml">
                 <field name="currency_id" position="after">
-                    <field name="typology_id" on_change="on_change_typology_id(typology_id)"/>
+                    <field name="type_id" on_change="on_change_type_id(type_id)"/>
                 </field>
             </field>
         </record>
